Enable button training off-the-shelf
If radios are in default P2P state, and training is pressed, units now complete training. If a different mode is used, or if server is set, server will stay in training until button press or command is issued. If a user wants to train specific custom settings in P2P mode using the training button, the user should put the unit with the desired settings into training first (ie have its train button pressed first).