Fix the macro that sets the ackTimer
Fix bug when ackTimer is zero and a comparison is done immediately.
millis() = 0
ackTimer = 1
       -  ___
          -1 = 0xFFFFFFFF
If the resulting value is treated as an unsigned number it is larger
than all of the timeout values.

Fix:
millis() = 0
ackTimer = -1 (0xFFFFFFFF)
        - ____
           1

This result value reduces the expected timeout value by one millisecond
which is significantly less than all of the timeouts.