Define a single source for channelTimer corrections
For point-to-point mode the source for channel timer corrections is the
system that sent the DATAGRAM_SYNC_CLOCKS frame.  The other system is
the only system that updates its timer.

For multi-point and virtual-circuit modes the server is the source for
channel timer corrections.  The client systems all synchronize to the
server.