Start hopping channels with TX/RX of SYNC_CLOCKS
Set the initial clock synchronization at less than a millisecond.  The
separation of the TX and RX transactionComplete interrupts is around 650
- 700 uSec.  By letting the transactionCompleteISR routine call
startChannelTimer, the first expire time of the two timers is within 1
millisecond.

The transmit completion routine for SYNC_CLOCKS and the receive routine
for SYNC_CLOCKS both call hopChannel to change to the next frequency.
This ensures that the two radios start hopping as soon as they are
synchronized.