Quick Start: Add references to command mode
Mention that command mode can be used to change the network ID and
encryption key.

Add links to the other documentation.

------

Note that the various web browsers behave differently.  While Microsoft
Edge might display the menu as a list of items on the left, Google
Chrome only displays a menu icon and unless the user knows that there is
a menu, the icon is easy to miss.

As such, I recommend adding a short paragraphs describing the other
sections of the documentation and include a link in the paragraph.  This
way if the user does not notice the menu button they still have access
to the full documentation.