New training button behavior
When the training button is pressed the green LEDs start counting
seconds.  When the 4 second boundary is crossed the yellow LED starts to
blink.  When the 8 second boundary is crossed the blue LED starts to
blink.  When the time reaches 15 seconds the radio clears NVM and
resets.  Upon boot the LoRaSerial radio starts with factory settings.